separated hours_to_time to utils folder
