+++ freqtrade/exchange/ftx.py
@@ -6,10 +6,10 @@ import ccxt
 from datetime import time
 from freqtrade.exceptions import (DDosProtection, InsufficientFundsError, InvalidOrderException,
                                   OperationalException, TemporaryError)
-from freqtrade.exchange import Exchange, hours_to_time
+from freqtrade.exchange import Exchange
 from freqtrade.exchange.common import API_FETCH_ORDER_RETRY_COUNT, retrier
 from freqtrade.misc import safe_value_fallback2
-
+from freqtrade.utils import hours_to_time
 
 logger = logging.getLogger(__name__)
