+++ freqtrade/utils/__init__.py
@@ -1,0 +1,2 @@
+# flake8: noqa: F401
+from freqtrade.utils.hours_to_time import hours_to_time
